refactor: update Verify to new approach
