+++ koala/cogs/__init__.py
@@ -6,3 +6,4 @@ from .colour_role import ColourRole
 from .intro_cog import IntroCog
 from .react_for_role import ReactForRole
 from .text_filter import TextFilter
+from .verification import Verification
