+++ koala/cogs/verification/cog.py
@@ -1,0 +1,431 @@
+#!/usr/bin/env python
+
+"""
+Koala Bot Base Cog code and additional base cog functions
+Commented using reStructuredText (reST)
+"""
+# Futures
+
+# Built-in/Generic Imports
+import random
+import string
+import smtplib
+from email.message import EmailMessage
+from dotenv import load_dotenv
+import os
+
+# Libs
+import discord
+from discord.ext import commands
+from sqlalchemy import select, delete, or_, and_, text
+
+# Own modules
+import KoalaBot
+from koala.db import session_manager
+
+from .models import VerifiedEmails, NonVerifiedEmails, Roles, ToReVerify
+from .env import GMAIL_EMAIL, GMAIL_PASSWORD
+from .log import logger
+
+# Constants
+
+# Variables
+
+
+
+
+def verify_is_enabled(ctx):
+    """
+    A command used to check if the guild has enabled verify
+    e.g. @commands.check(verify_is_enabled)
+    :param ctx: The context of the message
+    :return: True if enabled or test, False otherwise
+    """
+    try:
+        result = KoalaBot.check_guild_has_ext(ctx, "Verify")
+    except PermissionError:
+        result = False
+
+    return result or (str(ctx.author) == KoalaBot.TEST_USER and KoalaBot.is_dpytest)
+
+
+# FIXME: Move database commands to db.py
+class Verification(commands.Cog, name="Verify"):
+
+    def __init__(self, bot):
+        self.bot = bot
+        KoalaBot.database_manager.insert_extension("Verify", 0, True, True)
+
+    @staticmethod
+    def send_email(email, token):
+        """
+        Sends an email through gmails smtp server from the email stored in the environment variables
+        :param email: target to send an email to
+        :param token: the token the recipient will need to verify with
+        :return:
+        """
+        email_server = smtplib.SMTP_SSL('smtp.gmail.com', 465)
+        email_server.ehlo()
+        username = GMAIL_EMAIL
+        password = GMAIL_PASSWORD
+
+        msg = EmailMessage()
+        msg.set_content(f"Please send the bot the command:\n\n{KoalaBot.COMMAND_PREFIX}confirm {token}")
+        msg['Subject'] = "Koalabot Verification"
+        msg['From'] = username
+        msg['To'] = email
+
+        email_server.login(username, password)
+        email_server.send_message(msg)
+        email_server.quit()
+
+    @commands.Cog.listener()
+    async def on_ready(self):
+        await self.assign_roles_on_startup()
+
+    @commands.Cog.listener()
+    async def on_member_join(self, member):
+        """
+        Assigns necessary roles to users upon joining a server
+        :param member: the member object who just joined a server
+        :return:
+        """
+        with session_manager() as session:
+            potential_emails = session.execute(select(Roles.r_id, Roles.email_suffix)
+                                               .filter_by(s_id=member.guild.id)).all()
+
+            if potential_emails:
+                roles = {}
+                for role_id, suffix in potential_emails:
+                    role = discord.utils.get(member.guild.roles, id=role_id)
+                    roles[suffix] = role
+                    results = session.execute(select(VerifiedEmails).where(
+                        and_(
+                            VerifiedEmails.email.endswith(suffix),
+                            VerifiedEmails.u_id == member.id
+                        ))).all()
+
+                    blacklisted = session.execute(select(ToReVerify)
+                                                  .filter_by(r_id=role_id, u_id=member.id)).all()
+
+                    if results and not blacklisted:
+                        await member.add_roles(role)
+                message_string = f"""Welcome to {member.guild.name}. This guild has verification enabled.
+Please verify one of the following emails to get the appropriate role using `{KoalaBot.COMMAND_PREFIX}verify your_email@example.com`.
+This email is stored so you don't need to verify it multiple times across servers."""
+                await member.send(
+                    content=message_string + "\n" + "\n".join([f"`{x}` for `@{y}`" for x, y in roles.items()]))
+
+    @commands.check(KoalaBot.is_admin)
+    @commands.command(name="verifyAdd", aliases=["addVerification"])
+    @commands.check(verify_is_enabled)
+    async def enable_verification(self, ctx, suffix=None, role=None):
+        """
+        Set up a role and email pair for KoalaBot to verify users with
+        :param ctx: context of the discord message
+        :param suffix: end of the email (e.g. "example.com")
+        :param role: the role to give users with that email verified (e.g. @students)
+        :return:
+        """
+        with session_manager() as session:
+            if not role or not suffix:
+                raise self.InvalidArgumentError(f"Please provide the correct arguments\n(`{KoalaBot.COMMAND_PREFIX}enable_verification <domain> <@role>`")
+
+            try:
+                role_id = int(role[3:-1])
+            except ValueError:
+                raise self.InvalidArgumentError("Please give a role by @mentioning it")
+            except TypeError:
+                raise self.InvalidArgumentError("Please give a role by @mentioning it")
+
+            role_valid = discord.utils.get(ctx.guild.roles, id=role_id)
+            if not role_valid:
+                raise self.InvalidArgumentError("Please mention a role in this guild")
+
+            exists = session.execute(select(Roles)
+                                     .filter_by(s_id=ctx.guild.id, r_id=role_id, email_suffix=suffix)).all()
+            if exists:
+                raise self.VerifyError("Verification is already enabled for that role")
+
+            session.add(Roles(s_id=ctx.guild.id, r_id=role_id, email_suffix=suffix))
+            session.commit()
+
+            await ctx.send(f"Verification enabled for {role} for emails ending with `{suffix}`")
+            await self.assign_role_to_guild(ctx.guild, role_valid, suffix)
+
+    @commands.check(KoalaBot.is_admin)
+    @commands.command(name="verifyRemove", aliases=["removeVerification"])
+    @commands.check(verify_is_enabled)
+    async def disable_verification(self, ctx, suffix=None, role=None):
+        """
+        Disable an existing verification listener
+        :param ctx: context of the discord message
+        :param suffix: end of the email (e.g. "example.com")
+        :param role: the role paired with the email (e.g. @students)
+        :return:
+        """
+        with session_manager() as session:
+            if not role or not suffix:
+                raise self.InvalidArgumentError(
+                    f"Please provide the correct arguments\n(`{KoalaBot.COMMAND_PREFIX}enable_verification <domain> <@role>`")
+
+            try:
+                role_id = int(role[3:-1])
+            except ValueError:
+                raise self.InvalidArgumentError("Please give a role by @mentioning it")
+            except TypeError:
+                raise self.InvalidArgumentError("Please give a role by @mentioning it")
+
+            session.execute(delete(Roles).filter_by(s_id=ctx.guild.id, r_id=role_id, email_suffix=suffix))
+            session.commit()
+
+            await ctx.send(f"Emails ending with {suffix} no longer give {role}")
+
+
+    @commands.check(KoalaBot.is_dm_channel)
+    @commands.command(name="verify")
+    async def verify(self, ctx, email):
+        """
+        Send to KoalaBot in dms to verify an email with our system
+        :param ctx: the context of the discord message
+        :param email: the email you want to verify
+        :return:
+        """
+        with session_manager() as session:
+            already_verified = session.execute(select(VerifiedEmails).filter_by(email=email)).all()
+
+            in_blacklist = session.execute(select(ToReVerify).filter_by(u_id=ctx.author.id)).all()
+
+            if already_verified and not in_blacklist:
+                raise self.VerifyError("That email is already verified")
+
+            verification_code = ''.join(random.choice(string.ascii_letters) for _ in range(8))
+            session.add(NonVerifiedEmails(u_id=ctx.author.id, email=email, token=verification_code))
+            session.commit()
+
+            self.send_email(email, verification_code)
+            await ctx.send("Please verify yourself using the command you have been emailed")
+
+    @commands.check(KoalaBot.is_dm_channel)
+    @commands.command(name="unVerify")
+    async def un_verify(self, ctx, email):
+        """
+        Send to KoalaBot in dms to un-verify an email with our system
+        :param ctx: the context of the discord message
+        :param email: the email you want to un-verify
+        :return:
+        """
+        with session_manager() as session:
+            entry = session.execute(select(VerifiedEmails).filter_by(u_id=ctx.author.id, email=email)).all()
+
+            if not entry:
+                raise self.VerifyError("You have not verified that email")
+
+            session.execute(delete(VerifiedEmails).filter_by(u_id=ctx.author.id, email=email))
+            session.commit()
+
+            await self.remove_roles_for_user(ctx.author.id, email)
+            await ctx.send(f"{email} has been un-verified and relevant roles have been removed")
+
+    @commands.check(KoalaBot.is_dm_channel)
+    @commands.command(name="confirm")
+    async def confirm(self, ctx, token):
+        """
+        Send to KoalaBot in dms to confirm the verification of an email
+        :param ctx: the context of the discord message
+        :param token: the token emailed to you to verify with
+        :return:
+        """
+        with session_manager() as session:
+            entry = session.execute(select(NonVerifiedEmails).filter_by(token=token)).scalar()
+
+            if not entry:
+                raise self.InvalidArgumentError("That is not a valid token")
+
+            already_verified = session.execute(select(VerifiedEmails)
+                                               .filter_by(u_id=ctx.author.id, email=entry.email)).all()
+
+            if not already_verified:
+                session.add(VerifiedEmails(u_id=ctx.author.id, email=entry.email))
+
+            session.execute(delete(NonVerifiedEmails).filter_by(token=token))
+
+            potential_roles = session.execute(select(Roles.r_id)
+                                              .where(text(":email like ('%' || email_suffix)")),
+                                              {"email": entry.email}).all()
+            if potential_roles:
+                for role_id in potential_roles:
+                    session.execute(delete(ToReVerify).filter_by(r_id=role_id[0], u_id=ctx.author.id))
+
+            session.commit()
+            await ctx.send("Your email has been verified, thank you")
+            await self.assign_roles_for_user(ctx.author.id, entry.email)
+
+    @commands.check(KoalaBot.is_owner)
+    @commands.command(name="getEmails")
+    async def get_emails(self, ctx, user_id: int):
+        """
+        See the emails a user is verified with
+        :param ctx: the context of the discord message
+        :param user_id: the id of the user who's emails you want to find
+        :return:
+        """
+        with session_manager() as session:
+            results = session.execute(select(VerifiedEmails.email).filter_by(u_id=user_id)).all()
+
+            emails = '\n'.join([x[0] for x in results])
+            await ctx.send(f"This user has registered with:\n{emails}")
+
+    @commands.command(name="verifyList", aliases=["checkVerifications"])
+    @commands.check(verify_is_enabled)
+    async def check_verifications(self, ctx):
+        """
+        List the current verification setup for the server
+        :param ctx: the context of the discord message
+        :return:
+        """
+        with session_manager() as session:
+            embed = discord.Embed(title=f"Current verification setup for {ctx.guild.name}")
+            roles = session.execute(select(Roles.r_id, Roles.email_suffix).filter_by(s_id=ctx.guild.id)).all()
+
+            role_dict = {}
+            for role_id, suffix in roles:
+                role = discord.utils.get(ctx.guild.roles, id=role_id)
+                try:
+                    if suffix in role_dict:
+                        role_dict[suffix].append("@" + role.name)
+                    else:
+                        role_dict[suffix] = ["@" + role.name]
+                except AttributeError as e:
+                    session.execute(delete(Roles).filter_by(r_id=role_id))
+
+            session.commit()
+            for suffix, roles in role_dict.items():
+                embed.add_field(name=suffix, value='\n'.join(roles))
+
+            await ctx.send(embed=embed)
+
+    @commands.check(KoalaBot.is_admin)
+    @commands.command(name="reVerify")
+    @commands.check(verify_is_enabled)
+    async def re_verify(self, ctx, role):
+        """
+        Removes a role from all users who have it and marks them as needing to re-verify before giving it back
+        :param ctx: the context of the discord message
+        :param role: the role to be removed and re-verified (e.g. @students)
+        :return:
+        """
+        with session_manager() as session:
+            try:
+                role_id = int(role[3:-1])
+            except ValueError:
+                raise self.InvalidArgumentError("Please give a role by @mentioning it")
+            except TypeError:
+                raise self.InvalidArgumentError("Please give a role by @mentioning it")
+
+            exists = session.execute(select(Roles).filter_by(s_id=ctx.guild.id, r_id=role_id)).all()
+
+            if not exists:
+                raise self.VerifyError("Verification is not enabled for that role")
+            role = discord.utils.get(ctx.guild.roles, id=role_id)
+            for member in ctx.guild.members:
+                if role in member.roles:
+                    await member.remove_roles(role)
+                    session.add(ToReVerify(u_id=member.id, r_id=role.id))
+
+            session.commit()
+            await ctx.send("That role has now been removed from all users and they will need to re-verify the associated email.")
+
+    class InvalidArgumentError(Exception):
+        pass
+
+    class VerifyError(Exception):
+        pass
+
+    async def assign_roles_on_startup(self):
+        with session_manager() as session:
+            results = session.execute(select(Roles.s_id, Roles.r_id, Roles.email_suffix)).all()
+            for g_id, r_id, suffix in results:
+                try:
+                    guild = self.bot.get_guild(g_id)
+                    role = discord.utils.get(guild.roles, id=r_id)
+                    await self.assign_role_to_guild(guild, role, suffix)
+                except AttributeError as e:
+                    # bot not in guild
+                    print(e)
+
+    async def assign_roles_for_user(self, user_id, email):
+        with session_manager() as session:
+            results = session.execute(select(Roles.s_id, Roles.r_id, Roles.email_suffix)
+                                      .where(text(":email like ('%' || email_suffix)")), {"email": email}).all()
+
+            for g_id, r_id, suffix in results:
+                blacklisted = session.execute(select(ToReVerify).filter_by(r_id=r_id, u_id=user_id)).all()
+
+                if blacklisted:
+                    continue
+                try:
+                    guild = self.bot.get_guild(g_id)
+                    role = discord.utils.get(guild.roles, id=r_id)
+                    member = guild.get_member(user_id)
+                    if not member:
+                        member = await guild.fetch_member(user_id)
+                    await member.add_roles(role)
+                except AttributeError as e:
+                    # bot not in guild
+                    print(e)
+                except discord.errors.NotFound:
+                    print(f"user with id {user_id} not found")
+
+    async def remove_roles_for_user(self, user_id, email):
+        with session_manager() as session:
+            results = session.execute(select(Roles.s_id, Roles.r_id, Roles.email_suffix)
+                                      .where(text(":email like ('%' || email_suffix)")), {"email": email}).all()
+
+            for g_id, r_id, suffix in results:
+                try:
+                    guild = self.bot.get_guild(g_id)
+                    role = discord.utils.get(guild.roles, id=r_id)
+                    member = guild.get_member(user_id)
+                    if not member:
+                        member = await guild.fetch_member(user_id)
+                    await member.remove_roles(role)
+                except AttributeError as e:
+                    # bot not in guild
+                    print(e)
+                except discord.errors.NotFound:
+                    print(f"user with id {user_id} not found in {guild}")
+
+    async def assign_role_to_guild(self, guild, role, suffix):
+        with session_manager() as session:
+            results = session.execute(select(VerifiedEmails.u_id).where(VerifiedEmails.email.endswith(suffix))).all()
+
+            for user_id in results:
+                try:
+                    blacklisted = session.execute(select(ToReVerify).filter_by(r_id=role.id, u_id=user_id[0])).all()
+
+                    if blacklisted:
+                        continue
+                    member = guild.get_member(user_id[0])
+                    if not member:
+                        member = await guild.fetch_member(user_id[0])
+                    await member.add_roles(role)
+                except AttributeError as e:
+                    # bot not in guild
+                    print(e)
+                except discord.errors.NotFound:
+                    print(f"user with id {user_id} not found in {guild}")
+
+
+def setup(bot: KoalaBot) -> None:
+    """
+    Load this cog to the KoalaBot.
+    :param bot: the bot client for KoalaBot
+    """
+    if GMAIL_EMAIL is None or GMAIL_PASSWORD is None:
+        logger.warning("Verification not started. API keys not found in environment.")
+        KoalaBot.database_manager.insert_extension("Verify", 0, False, False)
+    else:
+        bot.add_cog(Verification(bot))
+        logger.info("Verification is ready.")
+
