+++ koala/cogs/verification/env.py
@@ -1,0 +1,7 @@
+import os
+from dotenv import load_dotenv
+
+load_dotenv()
+
+GMAIL_EMAIL = os.environ.get('GMAIL_EMAIL')
+GMAIL_PASSWORD = os.environ.get('GMAIL_PASSWORD')
